requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
Pillow==10.1.0
numpy==1.26.0
scikit-image==0.22.0
